main: change type in default flowgraph to float
The default types in the registry were changed from double to float,
this changes them in the builtin grc string to be able to launch without
parameters.

+++ src/service/main.cpp
@@ -71,12 +71,12 @@ int main(int argc, char** argv) {
     std::string grc = R"(blocks:
   - name: source
     id: gr::basic::ClockSource
-    template_args: double
+    template_args: float
   - name: sink
     id: gr::basic::DataSink
     parameters:
       signal_name: test
-    template_args: double
+    template_args: float
 connections:
   - [source, 0, sink, 0]
 )";
